@@ -6,34 +6,24 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
-import android.graphics.Bitmap;
-import android.graphics.Canvas;
-import android.graphics.Color;
-import android.graphics.Paint;
 import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.support.v4.view.ViewPager;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.viewpager.widget.ViewPager;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ImageButton;
 import android.widget.RadioButton;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.android.volley.Request;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.JsonArrayRequest;
-import com.android.volley.toolbox.JsonObjectRequest;
 import com.anmol.rosei.Adapter.GridAdapter;
 import com.anmol.rosei.Adapter.ViewpageAdapter;
 import com.anmol.rosei.Helpers.AuthUser;
@@ -56,11 +46,6 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-import org.w3c.dom.Text;
-
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -73,7 +58,7 @@ import de.hdodenhof.circleimageview.CircleImageView;
 public class RoseiActivity extends AppCompatActivity implements SheetLayout.OnFabAnimationEndListener{
 
     Animation rotate;
-    android.support.design.widget.FloatingActionButton book;
+    com.google.android.material.floatingactionbutton.FloatingActionButton book;
     TextView stuid;
     TextView user;
     private static long back_pressed;
@@ -97,7 +82,7 @@ public class RoseiActivity extends AppCompatActivity implements SheetLayout.OnFa
         setContentView(R.layout.activity_rosei);
         rotate = AnimationUtils.loadAnimation(RoseiActivity.this,R.anim.rotate);
         stuimg = (CircleImageView)findViewById(R.id.stuimg);
-        book = (android.support.design.widget.FloatingActionButton) findViewById(R.id.book);
+        book = (com.google.android.material.floatingactionbutton.FloatingActionButton) findViewById(R.id.book);
         about = (FloatingActionButton)findViewById(R.id.pp);
         epi = (FloatingActionButton)findViewById(R.id.epi);
         cd = (FloatingActionButton)findViewById(R.id.cd);
